+++ docs/change_log.md
@@ -1,6 +1,22 @@
 # Changelog
 All notable changes to this project will be documented in this file. We keep track of changes in this file since v0.1.8. The source code for most releases is available on [GitHub](https://github.com/mikekeith52/scalecast).
 
+## [0.14.8] - 2022-10-14
+### Added
+- added `must_keep` arg to `Forecater.auto_Xvar_select()`.
+- added `SeriesTransformer.SqrtTransform()` and `SeriesTransformer.SqrtRevert()` functions.
+### Changed
+- two-level differencing no longer natively supported in `Forecaster` and `MVForecaster`. it's too much work to maintain two differencing and it is also supported more efficiently and dynamically through the `SeriesTransformer` object and that is available for who needs it.
+  - this changed the arguments in the following functions:
+    - [`Forecaster.diff()`](https://scalecast.readthedocs.io/en/latest/Forecaster/Forecaster.html#src.scalecast.Forecaster.Forecaster.diff)
+    - [`Forecaster.integrate()`](https://scalecast.readthedocs.io/en/latest/Forecaster/Forecaster.html#src.scalecast.Forecaster.Forecaster.integrate)
+    - [`Forecaster.undiff()`](https://scalecast.readthedocs.io/en/latest/Forecaster/Forecaster.html#src.scalecast.Forecaster.Forecaster.undiff)
+    - [`util.pdr_load()`](https://scalecast.readthedocs.io/en/latest/Forecaster/Util.html#module-src.scalecast.util.pdr_load)
+- changed accepted values that can be passed to the `probabilistic` arg in the `Forecaster.tune_test_forecast()`, `MVForecaster.tune_test_forecast()`, and `notebook.tune_test_forecast()` functions so that some models can be forecasted probabilistically and others don't have to be (speeds up processing time generally).
+### Fixed
+- issues with second-differencing have all been resolved since second differencing is no longer supported
+- took out `revert_fvs` from `SeriesTransformer.DiffRevert()` function because it hasn't been working.
+
 ## [0.14.7] - 2022-10-05
 ### Added
 ### Changed